move slice_manual to peakgroup
