+++ alphadia/numba/fragments.py
@@ -9,8 +9,6 @@ import numba as nb
 import numpy as np
 from numba.extending import overload_method
 
-from alphadia.utils import USE_NUMBA_CACHING
-
 
 @nb.experimental.jitclass()
 class FragmentContainer:
@@ -202,59 +200,3 @@ def slice(inst, slices):
             return f
 
         return impl
-
-
-@nb.njit(cache=USE_NUMBA_CACHING)
-def slice_manual(inst, slices):
-    precursor_idx = []
-    fragments_mz_library = []
-    fragments_mz = []
-    fragments_intensity = []
-    fragments_type = []
-    fragments_loss_type = []
-    fragments_charge = []
-    fragments_number = []
-    fragments_position = []
-    fragments_cardinality = []
-
-    precursor = np.arange(len(slices), dtype=np.uint32)
-
-    for i, (start_idx, stop_idx, _step) in enumerate(slices):
-        for j in range(start_idx, stop_idx):
-            precursor_idx.append(precursor[i])
-            fragments_mz_library.append(inst.mz_library[j])
-            fragments_mz.append(inst.mz[j])
-            fragments_intensity.append(inst.intensity[j])
-            fragments_type.append(inst.type[j])
-            fragments_loss_type.append(inst.loss_type[j])
-            fragments_charge.append(inst.charge[j])
-            fragments_number.append(inst.number[j])
-            fragments_position.append(inst.position[j])
-            fragments_cardinality.append(inst.cardinality[j])
-
-    precursor_idx = np.array(precursor_idx, dtype=np.uint32)
-    fragments_mz_library = np.array(fragments_mz_library, dtype=np.float32)
-    fragment_mz = np.array(fragments_mz, dtype=np.float32)
-    fragment_intensity = np.array(fragments_intensity, dtype=np.float32)
-    fragment_type = np.array(fragments_type, dtype=np.uint8)
-    fragment_loss_type = np.array(fragments_loss_type, dtype=np.uint8)
-    fragment_charge = np.array(fragments_charge, dtype=np.uint8)
-    fragment_number = np.array(fragments_number, dtype=np.uint8)
-    fragment_position = np.array(fragments_position, dtype=np.uint8)
-    fragment_cardinality = np.array(fragments_cardinality, dtype=np.uint8)
-
-    f = FragmentContainer(
-        fragments_mz_library,
-        fragment_mz,
-        fragment_intensity,
-        fragment_type,
-        fragment_loss_type,
-        fragment_charge,
-        fragment_number,
-        fragment_position,
-        fragment_cardinality,
-    )
-
-    f.precursor_idx = precursor_idx
-
-    return f
